Use fewer ints

+++ geobuf.cpp
@@ -27,7 +27,7 @@ struct queued_feature {
 	double e = 0;
 	std::vector<std::string> *keys = NULL;
 	std::vector<struct serialization_state> *sst = NULL;
-	int layer = 0;
+	size_t layer = 0;
 	std::string layername = "";
 };
 
@@ -145,7 +145,7 @@ drawvec readLine(std::vector<long> &coords, size_t dim, double e, bool closed) {
 	return readLinePart(coords, dim, e, 0, coords.size(), closed);
 }
 
-drawvec readMultiLine(std::vector<long> &coords, std::vector<int> &lengths, size_t dim, double e, bool closed) {
+drawvec readMultiLine(std::vector<long> &coords, std::vector<size_t> &lengths, size_t dim, double e, bool closed) {
 	if (lengths.size() == 0) {
 		return readLinePart(coords, dim, e, 0, coords.size(), closed);
 	}
@@ -164,7 +164,7 @@ drawvec readMultiLine(std::vector<long> &coords, std::vector<int> &lengths, size
 	return dv;
 }
 
-drawvec readMultiPolygon(std::vector<long> &coords, std::vector<int> &lengths, size_t dim, double e) {
+drawvec readMultiPolygon(std::vector<long> &coords, std::vector<size_t> &lengths, size_t dim, double e) {
 	ensureDim(dim);
 
 	if (lengths.size() == 0) {
@@ -203,7 +203,7 @@ struct drawvec_type {
 std::vector<drawvec_type> readGeometry(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys) {
 	std::vector<drawvec_type> ret;
 	std::vector<long> coords;
-	std::vector<int> lengths;
+	std::vector<size_t> lengths;
 	int type = -1;
 
 	while (pbf.next()) {
@@ -265,7 +265,7 @@ std::vector<drawvec_type> readGeometry(protozero::pbf_reader &pbf, size_t dim, d
 	return ret;
 }
 
-void readFeature(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, struct serialization_state *sst, int layer, std::string layername) {
+void readFeature(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, struct serialization_state *sst, size_t layer, std::string layername) {
 	std::vector<drawvec_type> dv;
 	long id = 0;
 	bool has_id = false;
@@ -480,7 +480,7 @@ void runQueue() {
 	feature_queue.clear();
 }
 
-void queueFeature(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, std::vector<struct serialization_state> *sst, int layer, std::string layername) {
+void queueFeature(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, std::vector<struct serialization_state> *sst, size_t layer, std::string layername) {
 	struct queued_feature qf;
 	qf.pbf = pbf;
 	qf.dim = dim;
@@ -497,7 +497,7 @@ void queueFeature(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<
 	}
 }
 
-void outBareGeometry(drawvec const &dv, int type, struct serialization_state *sst, int layer, std::string layername) {
+void outBareGeometry(drawvec const &dv, int type, struct serialization_state *sst, size_t layer, std::string layername) {
 	serial_feature sf;
 
 	sf.layer = layer;
@@ -515,7 +515,7 @@ void outBareGeometry(drawvec const &dv, int type, struct serialization_state *ss
 	serialize_feature(sst, sf);
 }
 
-void readFeatureCollection(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, std::vector<struct serialization_state> *sst, int layer, std::string layername) {
+void readFeatureCollection(protozero::pbf_reader &pbf, size_t dim, double e, std::vector<std::string> &keys, std::vector<struct serialization_state> *sst, size_t layer, std::string layername) {
 	while (pbf.next()) {
 		switch (pbf.tag()) {
 		case 1: {
@@ -530,7 +530,7 @@ void readFeatureCollection(protozero::pbf_reader &pbf, size_t dim, double e, std
 	}
 }
 
-void parse_geobuf(std::vector<struct serialization_state> *sst, const char *src, size_t len, int layer, std::string layername) {
+void parse_geobuf(std::vector<struct serialization_state> *sst, const char *src, size_t len, size_t layer, std::string layername) {
 	protozero::pbf_reader pbf(src, len);
 
 	size_t dim = 2;
